Antflip 1.7.0
Set PSWAP based on antenna state from K3. The user may override it
until the band or antenna is changed again. The configuration format
has been updated to support:

    pswap = [true, false]

where the first value is for antenna 1 and the second value is for
antenna 2. A single value will be used for antenna 1 and its inverse
for antenna 2.