Support WinK3Suite for K3 communications.
Send serial commands to K3 every 200ms. There is no queueing and
commands are sent blindly.

Update UI to show serial port state without popups and support
reconnecting serial port. Only reconnect on explicit user command and
not when selecting the COM port. Expose baud rate even though it should
not matter for virtual serial ports.

Enforce a delay between closing the previous serial connection and
starting the next to avoid UI hangs and failed connects in
System.IO.Ports which uses a thread pool under the covers.  Serialize
access to the serial port loop to avoid issues if a user presses the
reconnect button too fast.

Log messages (poorly) to the application event log. Try to avoid
crashing the application if there is an issue with the COM port.